# Conflicts:
#	app/service/file.go
#	go.mod

+++ app/common/type.go
@@ -14,8 +14,10 @@ func (d Date) MarshalJSON() ([]byte, error) {
 }
 
 // BaseModel soft_delete 此插件还支持混合模式，删除同时更新删除时间
+//
+// 主键全部采用雪花算法id
 type BaseModel struct {
-	ID        uint                  `gorm:"primaryKey" json:"id"`
+	ID        uint64                `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time             `json:"createdAt"`
 	UpdatedAt time.Time             `json:"-"`
 	IsDel     soft_delete.DeletedAt `gorm:"softDelete:flag;not null;default:0" json:"-"`
