+++ app/controller/userController/userIpVo/userIpPageReqVo.go
@@ -8,7 +8,7 @@ import (
 type UserIpPageReqVo struct {
 	common.PageReq
 	Ip        string    `form:"ip" json:"ip"`
-	UserId    uint      `form:"userId" json:"userId"`
+	UserId    uint64    `form:"userId" json:"userId"`
 	StartTime time.Time `form:"startTime" json:"startTime"`
 	EndTime   time.Time `form:"endTime" json:"endTime"`
 }
