+++ app/controller/userController/userIpVo/userIpRespVo.go
@@ -6,8 +6,8 @@ import (
 )
 
 type UserIpRespVo struct {
-	ID        uint        `gorm:"primaryKey" json:"id"`
-	UserId    uint        `gorm:"not null;comment:用户id" json:"userId"`
+	ID        uint64      `gorm:"primaryKey" json:"id"`
+	UserId    uint64      `gorm:"not null;comment:用户id" json:"userId"`
 	Ip        string      `gorm:"size:45;comment:ip地址" json:"ip"`
 	Country   string      `gorm:"size:30;comment:国家" json:"country"`
 	Province  string      `gorm:"size:30;comment:省份" json:"province"`
